honksquad: refactor: Illiterate -> Papyrophobia, scope to paper only
Rescopes the quirk from a broad illiteracy fantasy to what the
implementation actually does: the holder can't interact with paper.
Cost drops from -8 to -2 since computer consoles, PDAs, and signage
remain fully usable.

- Component and system renamed and moved to the @RussStation fork dir
- Popup and locale keys reworded as a phobia response
- BlockWriting dropped from the trait and clone list; the write block
  is handled directly in PapyrophobiaSystem now
- Exclusion tag changed from "literacy" to "paper"
