+++ Content.Shared/RussStation/Traits/PapyrophobiaComponent.cs
@@ -1,0 +1,6 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.RussStation.Traits;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class PapyrophobiaComponent : Component;
